Write json file to own file, add template for port
A json file with all of the portfolio info is now written to the disk for
use as data to load in with React.

React router uses params to determine which child to render instead of the
router itself. Pass in the right info for each page based on which child
is to render.

Begin styling of component and organization of content.

Remove excess files.

install json-loader via npm.